+++ app/assets/javascripts/admin/api_flow_types.js
@@ -218,6 +218,7 @@ type APIProjectTypeBase = {
   +priority: number,
   +paused: boolean,
   +expectedTime: number,
+  +isBlacklistedFromReport: boolean,
 };
 
 export type APIProject = APIProjectTypeBase & {