+++ conf/evolutions/reversions/030-tracingstore.sql
@@ -3,6 +3,6 @@ START TRANSACTION;
 DROP VIEW webknossos.tracingStores_;
 DROP TABLE webknossos.tracingStores;
 
-UPDATE webknossos.releaseInformation SET schemaVersion = 27;
+UPDATE webknossos.releaseInformation SET schemaVersion = 29;
 
 COMMIT TRANSACTION;