+++ project/AssetCompilation.scala
@@ -1,7 +1,9 @@
 import java.nio.file.{Files, StandardCopyOption}
+import java.util.function.Consumer
 
-import sbt.Keys.{runner, _}
-import sbt.{Task, _}
+import sbt.Keys._
+import sbt._
+import sys.process.Process
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -33,7 +35,7 @@ object AssetCompilation {
 
   private def npmInstall: Def.Initialize[Task[Seq[File]]] = Def task {
     try{
-      val exitValue = startProcess(npmPath.value, List("install"), baseDirectory.value) ! streams.value.log
+      val exitValue = startProcess(npmPath.value, List("install", "--frozen-lockfile"), baseDirectory.value) ! streams.value.log
       if(exitValue != 0)
         throw new Error(s"Running npm failed with exit code: $exitValue")
     } catch {
@@ -65,7 +67,27 @@ object AssetCompilation {
     }
   }
 
+  private def copyRecursively(from: File, to: File): Unit = {
+    def toConsumer[A](function: A => Any): Consumer[A] = new Consumer[A]() {
+      override def accept(arg: A): Unit = function.apply(arg)
+    }
+
+    Files.walk(from.toPath).forEach(toConsumer(cpSrc => {
+      val cpDest = to.toPath.resolve(from.toPath.relativize(cpSrc))
+      Files.copy(cpSrc, cpDest)
+    }))
+  }
+
+
+  private def deleteRecursively(file: File): Unit = {
+    if (file.isDirectory)
+      file.listFiles.foreach(deleteRecursively)
+    if (!file.delete && file.exists)
+      throw new Exception(s"Unable to delete ${file.getAbsolutePath}")
+  }
+
   private def assetsGenerationTask: Def.Initialize[Task[Unit]] = Def task {
+    // copy tools/postgres
     try {
       val destination = target.value  / "universal" / "stage" / "tools" / "postgres"
       destination.mkdirs
@@ -75,6 +97,32 @@ object AssetCompilation {
     } catch {
       case e: Exception => streams.value.log.error("Could not copy SQL schema to stage dir: " + e.getMessage)
     }
+
+    // copy node_modules for diff_schema.js
+    {
+      val nodeModules = Seq(
+        "commander",
+        "randomstring",
+        "glob",
+        "rimraf",
+        "replace-in-file"
+      )
+      val nodeSrc = baseDirectory.value / "node_modules"
+      val nodeDest = target.value  / "universal" / "stage" / "node_modules"
+      val tmpPath = baseDirectory.value / "tmp"
+      val streamsValue = streams.value.log
+
+      tmpPath.mkdirs
+      startProcess(npmPath.value, List("init", "-y"), tmpPath) ! streamsValue
+      nodeModules.foreach(nodeModule => {
+        startProcess(npmPath.value, List("add", (nodeSrc / nodeModule).getAbsolutePath), tmpPath) ! streamsValue
+      })
+      deleteRecursively(nodeDest)
+      copyRecursively(tmpPath / "node_modules", nodeDest)
+      deleteRecursively(tmpPath)
+    }
+
+    // run webpack
     try{
       val exitValue = startProcess(webpackPath.value, List("--env.production"), baseDirectory.value) ! streams.value.log
       if(exitValue != 0)
@@ -86,25 +134,30 @@ object AssetCompilation {
   } dependsOn npmInstall
 
   private def slickClassesFromDBSchemaTask: Def.Initialize[Task[Seq[File]]] = Def task{
+      val streamsValue = streams.value
+      val baseDirectoryValue = baseDirectory.value
+      val dependencyClasspathValue = (dependencyClasspath in Compile).value
+      val runnerValue = (runner in Compile).value
+      val sourceManagedValue = sourceManaged.value
 
-      val schemaPath = baseDirectory.value / "tools" / "postgres" / "schema.sql"
-      val slickTablesOutPath = sourceManaged.value / "schema" / "com" / "scalableminds" / "webknossos" / "schema" / "Tables.scala"
+      val schemaPath = baseDirectoryValue / "tools" / "postgres" / "schema.sql"
+      val slickTablesOutPath = sourceManagedValue / "schema" / "com" / "scalableminds" / "webknossos" / "schema" / "Tables.scala"
 
       val shouldUpdate = !slickTablesOutPath.exists || slickTablesOutPath.lastModified < schemaPath.lastModified
 
       if (shouldUpdate) {
-        streams.value.log.info("Ensuring Postgres DB is running for Slick code generation...")
-        startProcess((baseDirectory.value / "tools" / "postgres" / "ensure_db.sh").toString, List(), baseDirectory.value)  ! streams.value.log
+        streamsValue.log.info("Ensuring Postgres DB is running for Slick code generation...")
+        startProcess((baseDirectoryValue / "tools" / "postgres" / "ensure_db.sh").toString, List(), baseDirectoryValue)  ! streamsValue.log
 
-        streams.value.log.info("Updating Slick SQL schema from local database...")
+        streamsValue.log.info("Updating Slick SQL schema from local database...")
 
-        (runner in Compile).value.run("slick.codegen.SourceCodeGenerator", (dependencyClasspath in Compile).value.files,
-          Array("file://" + (baseDirectory.value / "conf" / "application.conf").toString + "#slick", (sourceManaged.value / "schema").toString),
-          streams.value.log
+        runnerValue.run("slick.codegen.SourceCodeGenerator", dependencyClasspathValue.files,
+          Array("file://" + (baseDirectoryValue / "conf" / "application.conf").toString + "#slick", (sourceManagedValue / "schema").toString),
+          streamsValue.log
         )
 
       } else {
-        streams.value.log.info("Slick SQL schema already up to date.")
+        streamsValue.log.info("Slick SQL schema already up to date.")
       }
 
       Seq((slickTablesOutPath))