+++ webknossos-tracingstore/app/com/scalableminds/webknossos/tracingstore/tracings/TracingDataStore.scala
@@ -1,10 +1,14 @@
 package com.scalableminds.webknossos.tracingstore.tracings
 
 import com.google.inject.Inject
-import com.scalableminds.webknossos.datastore.DataStoreConfig
 import com.scalableminds.webknossos.tracingstore.TracingStoreConfig
+import com.typesafe.scalalogging.LazyLogging
+import play.api.inject.ApplicationLifecycle
 
-class TracingDataStore @Inject()(config: TracingStoreConfig) {
+import scala.concurrent.Future
+
+class TracingDataStore @Inject()(config: TracingStoreConfig,
+                                 lifecycle: ApplicationLifecycle) extends LazyLogging {
 
   val healthClient = new FossilDBClient("healthCheckOnly", config)
 
@@ -18,4 +22,22 @@ class TracingDataStore @Inject()(config: TracingStoreConfig) {
 
   lazy val volumeData = new FossilDBClient("volumeData", config)
 
+  lazy val volumeUpdates = new FossilDBClient("volumeUpdates", config)
+
+  def shutdown() = {
+    healthClient.shutdown()
+    skeletons.shutdown()
+    skeletonUpdates.shutdown()
+    volumes.shutdown()
+    volumeData.shutdown()
+    volumeUpdates.shutdown()
+    ()
+  }
+
+  lifecycle.addStopHook { () =>
+    Future.successful {
+      logger.info("Closing TracingStore grpc channels...")
+      shutdown()
+    }
+  }
 }